% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initProcess}
\alias{initProcess}
\title{Process main streaming data for R}
\usage{
initProcess(fileName)
}
\arguments{
\item{fileName}{directory leading to the json file}
}
\description{
Process the main json file to produce a data frame where each
row states the song name, artist, length of the song played and the time it
was played
}
