% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot a time series of your data}
\usage{
plotTimeSeries(data, type = "hours")
}
\arguments{
\item{data}{the data produced from the initProcess fct}

\item{type}{specify on wat times scale, currently only hours}
}
\description{
Produces a radar chart/histogram detailing hours of the day
the user listens to music
}
