% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scrapeLocations}
\alias{scrapeLocations}
\title{Scrape artists country of origin}
\usage{
scrapeLocations(artistNames)
}
\arguments{
\item{artistNames}{a vector of artist names}
}
\description{
scrape the country of origin of each artist using wiki and 
last.fm
}
